user/requirements.txt
sqlalchemy
flask_cors
flask
flask_bcrypt
flask_sqlalchemy
pyjwt
gunicorn